[Iluvatar][SWCOMP-2466] Add Iluvatar ixcc backend port with vectorAdd support
- Switch MLIR pipeline from ROCDL to IXDL target for Iluvatar hardware
- Stub out emitMfma in FlyToROCDL.cpp (not needed for vectorAdd)
- Add FlyCorexRuntimeWrappers.cpp using CUDA Driver API for Iluvatar COREX
- Add conditional CMake build for COREX/ROCm runtime selection
- Fix FlyOps.cpp diagnostic API for ixcc compatibility
- Add comprehensive porting guide (docs/iluvatar-port-guide.md)

Made-with: Cursor

+++ lib/Dialect/Fly/IR/FlyOps.cpp
@@ -1611,9 +1611,8 @@ FLY_INFER_RETURN_TYPES(MmaMakeFragmentOp) {
     elemTy = mmaAtom.getValTypeD();
     break;
   default:
-    return emitOptionalError(location,
-                             "MmaMakeFragmentOp: invalid operand_id value: ",
-                             operandId);
+    return emitOptionalError(
+        location, "MmaMakeFragmentOp: invalid operand_id value: ", static_cast<int>(operandId));
   }
 
   LayoutAttr inputLayout = GetLayoutAttrFromLayoutLikeType(memrefTy);
